Add support for using local PBF files as data source
Previously, the repository downloaded during update all
files from the network. While this is still the default
behaviour, it is now possible to use local PBF files
as data source for repository update.

The repository will crawl the provided source directory
(by default planet/split) and use all PBF files it will
find in it (including files in all subdirs).

The path to the PBF file in the will be used as output
path in the results directory. So if you have:
europe/great-britain/england/kent.osm.pbf
The results from this PBF file will be stored in:
<results dir>/<data type>/europe/great-britain/england/kent